feat: mesh class need to implement objloader

+++ src/platform/Window.cpp
@@ -55,7 +55,7 @@ void    Window::close()
 
 const Input&    Window::pollEvents()
 {
-    _input._aspect = _width / _height;
+    _input._aspect = static_cast<double>(_width) / static_cast<double>(_height);
     _input._delta = _time.get() - lastFrameTime;
     lastFrameTime = _time.get();
 	SDL_Event   event;
